requirements.txt
fastapi
uvicorn
pytesseract
pdf2image
pillow
#easyocr
requests
python-multipart
faker